Add RemindStore::delete method
Currently, we're using build-in LinkedList implementation which is
a doubly linked list. Problem is when we're removing any element
from this list, it is taking O(n) operation to remove it. Make
sense as it doesn't provide us the underlying Node so it had to
iterate through the whole list when given an index or object.

Later, we'll implement our own doubly linked-list to make the
removal an O(1) operation.