Remove UPDATE operation
User only can send 3 types of info with RemindRequest:
1. context
	context can be updated but because context holds the info
that may relevant to a specific scenario, changing it would mean
entirely different scenario. So it's better be a new request.
2. schedules
	when an update comes for schedules, app may already be
processed some schedule expressions. So, it would be difficult
to reconcile with old schedule's data. Moreover, we expect most
of the scheduling will be one-time reminder. So, it would be far
more effective strategy to just delete it and create a new one.
Though for recurring reminder, it would be a hassle. But for now,
this design seems more simpler without compromising too much.
3. priority
	it's pretty trivial field. Though it seems like a good
idea to make some reminder requests more important than others
on the fly, but for now, we won't allow any update for priority
either. Maybe later, we can allow to update priority.

So, for now, no UPDATE operation.